Reorder components
