+++ src/components/library/artist/List.jsx
@@ -2,9 +2,9 @@ import React, { Component } from 'react'
 import { connect } from 'react-redux'
 
 import { loadLibraryEntries } from 'actions/library'
+import { withSearchParams } from 'components/adapted/ReactRouterDom'
 import ListingFetchWrapper from 'components/generics/ListingFetchWrapper'
 import Navigator from 'components/generics/Navigator'
-import { withSearchParams } from 'components/generics/Router'
 import ArtistEntry from 'components/library/artist/Entry'
 import SearchBox from 'components/library/SearchBox'
 import { artistStatePropType } from 'reducers/library'
