+++ src/components/library/song/Entry.jsx
@@ -6,9 +6,9 @@ import { CSSTransition, TransitionGroup } from 'react-transition-group'
 
 import { clearAlteration } from 'actions/alterations'
 import { addSongToPlaylist } from 'actions/playlist'
+import { withSearchParams } from 'components/adapted/ReactRouterDom'
+import { CSSTransitionLazy } from 'components/adapted/ReactTransitionGroup'
 import Notification from 'components/generics/Notification'
-import { CSSTransitionLazy } from 'components/generics/ReactTransitionGroup'
-import { withSearchParams } from 'components/generics/Router'
 import SongEntryExpanded from 'components/library/song/EntryExpanded'
 import { CanAddToPlaylist, IsPlaylistUser} from 'components/permissions/Playlist'
 import PlayQueueInfo from 'components/song/PlayQueueInfo'
