+++ src/components/library/song/EntryExpanded.jsx
@@ -5,9 +5,9 @@ import { connect } from 'react-redux'
 
 import { clearAlteration } from 'actions/alterations'
 import { addSongToPlaylistWithOptions } from 'actions/playlist'
+import { withSearchParams } from 'components/adapted/ReactRouterDom'
 import HighlighterQuery from 'components/generics/HighlighterQuery'
 import Notification from 'components/generics/Notification'
-import { withSearchParams } from 'components/generics/Router'
 import SongEntryExpandedArtist from 'components/library/song/EntryExpandedArtist'
 import SongEntryExpandedWork from 'components/library/song/EntryExpandedWork'
 import { CanAddToPlaylist, IsPlaylistUser} from 'components/permissions/Playlist'
