+++ src/components/library/song/List.jsx
@@ -4,9 +4,9 @@ import React, { Component } from 'react'
 import { connect } from 'react-redux'
 
 import { loadLibraryEntries } from 'actions/library'
+import { withSearchParams } from 'components/adapted/ReactRouterDom'
 import ListingFetchWrapper from 'components/generics/ListingFetchWrapper'
 import Navigator from 'components/generics/Navigator'
-import { withSearchParams } from 'components/generics/Router'
 import SearchBox from 'components/library/SearchBox'
 import SongEntry from 'components/library/song/Entry'
 import { songStatePropType } from 'reducers/library'
