+++ src/components/library/work/Entry.jsx
@@ -2,8 +2,8 @@ import PropTypes from 'prop-types'
 import { stringify } from 'query-string'
 import React, { Component } from 'react'
 
+import { withNavigate } from 'components/adapted/ReactRouterDom'
 import HighlighterQuery from 'components/generics/HighlighterQuery'
-import { withNavigate } from 'components/generics/Router'
 import { workPropType } from 'serverPropTypes/library'
 
 class WorkEntry extends Component {
