+++ src/components/settings/Tokens.jsx
@@ -5,9 +5,9 @@ import { CSSTransition } from 'react-transition-group'
 
 import { createPlayerToken, loadPlayerToken, revokePlayerToken } from 'actions/playlist'
 import { revokeToken } from 'actions/token'
+import { CSSTransitionLazy } from 'components/adapted/ReactTransitionGroup'
 import ConfirmationBar from 'components/generics/ConfirmationBar'
 import Notification from 'components/generics/Notification'
-import { CSSTransitionLazy } from 'components/generics/ReactTransitionGroup'
 import TokenWidget from 'components/generics/TokenWidget'
 import { IsLibraryManager } from 'components/permissions/Library'
 import { IsPlaylistManager } from 'components/permissions/Playlist'
