+++ src/components/settings/songTags/Entry.jsx
@@ -2,9 +2,9 @@ import classNames from 'classnames'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 
+import { CSSTransitionLazy } from 'components/adapted/ReactTransitionGroup'
 import { CheckboxField, FormInline, HueField } from 'components/generics/Form'
 import Notification, { NotifiableForTable } from 'components/generics/Notification'
-import { CSSTransitionLazy } from 'components/generics/ReactTransitionGroup'
 import { IsLibraryManager } from 'components/permissions/Library'
 import { alterationResponsePropType, Status } from 'reducers/alterationsResponse'
 import { songTagPropType } from 'serverPropTypes/library'
