+++ src/components/settings/songTags/List.jsx
@@ -5,9 +5,9 @@ import { connect } from 'react-redux'
 
 import { clearAlteration } from 'actions/alterations'
 import { editSongTag, getSongTagList } from 'actions/songTags'
+import { withLocation } from 'components/adapted/ReactRouterDom'
 import ListingFetchWrapper from 'components/generics/ListingFetchWrapper'
 import Navigator from 'components/generics/Navigator'
-import { withLocation } from 'components/generics/Router'
 import SettingsSongTagsEntry from 'components/settings/songTags/Entry'
 import { alterationResponsePropType } from 'reducers/alterationsResponse'
 import { songTagsStatePropType } from 'reducers/songTags'
